pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lora-trainer"
version = "1.0.0"
description = "LoRA Training Application - Dataset Generation and Model Fine-tuning"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = ["lora", "machine-learning", "fine-tuning", "nlp", "transformers"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "torch>=2.0.0",
    "transformers>=4.35.0",
    "tokenizers>=0.15.0",
    "peft>=0.7.0",
    "trl>=0.7.0",
    "accelerate>=0.25.0",
    "bitsandbytes>=0.41.0",
    "datasets>=2.14.0",
    "requests>=2.31.0",
    "numpy>=1.24.0",
    "pyyaml>=6.0",
    "matplotlib>=3.7.0",
    "gguf>=0.17.0",
    "tiktoken>=0.12.0",
    "protobuf>=6.0.0",
    "typer>=0.20.0",
    "rich>=14.0.0",
    "sentencepiece>=0.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[project.scripts]
lora-trainer = "lora_trainer_app.cli:main"

[project.urls]
Homepage = "https://github.com/yourusername/lora-trainer"
Documentation = "https://github.com/yourusername/lora-trainer#readme"
Repository = "https://github.com/yourusername/lora-trainer"

[tool.setuptools]
packages = ["lora_trainer_app"]

[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311']

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

